Moving the sample-secure-oauth2-actuator sample over to the integration-tests package
