+++ spring-boot-integration-tests/spring-boot-actuator-tests/spring-boot-actuator-test-secure-oauth2/src/main/resources/application.properties
