+++ spring-boot-integration-tests/spring-boot-actuator-tests/spring-boot-actuator-test-secure-oauth2/src/test/java/sample/SampleSecureOAuth2ActuatorApplicationTests.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sample.secure.oauth2.actuator;
+package sample;
 
 import org.junit.Before;
 import org.junit.Test;
